Update to use new plugin init mechanism
